Random placement
This may not be the perferred solution, because it requires a predefined post area to randomly place images within, but its the only way to do it without knowing the size of the images involved before hand. We can only find out image height and width when the image is first loaded. Saving a reference to these locations could work, but then none of them would be laid out until all of them were loaded. In this approach, as they load, they get placed randomly. If the random positions happen to overlap, a new random position is assigned, and no exact overlapping is ensured.