Page Top Buffer
Added a buffer to the top of the page where the menu will always appear. You'd think just having the scroll up would capture that, but safari has a bounce at the top of the page, which registers as a scroll down when you reach the top of the page with some velocity. This hides the menu, not the result we want when a reader gets to the top of the page.

Removed the image pile layout function. Anther has it integrated in WP.

Made a seperate script file for integrating with WP. Includes Modernizr touch text, and menu interaction layout.